backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
langchain-groq==0.0.1
faiss-cpu==1.7.4
sentence-transformers==2.3.1
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.21
openpyxl==3.1.2
python-dotenv==1.0.1
cors-headers==0.1.0